app/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
structlog>=24.1.0
prometheus_client>=0.20.0
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-instrumentation-fastapi>=0.43b0
opentelemetry-exporter-otlp>=1.22.0
python-multipart>=0.0.9
psutil>=5.9.8
pydantic-settings>=2.2.1
sqlalchemy[asyncio]>=2.0.28
asyncpg>=0.29.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
pytest>=8.0.0
httpx>=0.27.0
